backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
Werkzeug==2.3.0
requests==2.31.0
geopy==2.3.0
APScheduler==3.10.1
python-dateutil==2.8.2

